Changed tic variable type to native Uint32 instead of long; this fixes a few issues in the Wii version.

+++ include/object.h
@@ -20,7 +20,7 @@ extern int currentLevel;
 //The current frame
 int frame = 0;
 //The last frame to recieve input
-long lastInputTime = 0;
+Uint32 lastInputTime = 0;
 Object* player;
 int lastMoveDir;
 //Whether or not a player has been placed yet
